[UPDATE] All files Utilisation des double plutot que float pour la précision de calcul
[UPDATE] Objet
Ajout des champs et methode pour determiner les caractéristique d'un objet ( couleur , indice de refraction etc... ) // TODO Eventuellement faire une classe "materiaux" plus tard

[UPDATE] ParentPointVector | Ray
Ajout de getter et setter manquant

[UPDATE] Vector
Ajout du produit scalaire

[UPDATE] Cylinder
Ajout d'une fonction d'intersection ( non terminé )

[ADD] Scene
Pour gerer tout les objets affiche

[ADD] RayTracer
Gerer la scene , le nombre de rayon etc...

[ADD] Main
Main du programme

[UPDATE] Makefile
Ajout des nouveaux fichiers a compile

+++ include/Rectangle.hpp
@@ -11,17 +11,17 @@
 class Rectangle : public Objet {
 private:
     Point origin_;
-    float width_;
-    float height_;
-    float depth_;
+    double width_;
+    double height_;
+    double depth_;
 
 public:
     Rectangle(void);
-    Rectangle(Point origin, float width, float height,float depth);
+    Rectangle(Point origin, double width, double height,double depth);
 
     ~Rectangle();
 
-    bool intersect(Ray &ray, float &dist);
+    bool intersect(Ray &ray, double &dist);
 
 };
